refactor: RacingView를 RacingInputView, RacingOutputView로 분리
- input 로직과 output 로직에 맞춰서 분리
