+++ src/main/java/racingcar/view/RacingInputView.java
@@ -1,0 +1,19 @@
+package racingcar.view;
+
+import java.util.Scanner;
+
+public class RacingInputView {
+
+    private final Scanner scanner = new Scanner(System.in);
+
+
+    public String inputNames() {
+        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,)를 기준으로 구분) ");
+        return scanner.nextLine();
+    }
+
+    public String inputTrialNumber() {
+        System.out.println("시도할 회수는 몇회인가요?");
+        return scanner.nextLine();
+    }
+}
