+++ src/main/java/racingcar/view/RacingOutputView.java
@@ -4,23 +4,10 @@ import racingcar.model.Car;
 import racingcar.model.RacingCar;
 
 import java.util.List;
-import java.util.Scanner;
 import java.util.stream.Collectors;
 
-public class RacingView {
+public class RacingOutputView {
 
-    private final Scanner scanner = new Scanner(System.in);
-
-
-    public String inputNames() {
-        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,)를 기준으로 구분) ");
-        return scanner.nextLine();
-    }
-
-    public String inputTrialNumber() {
-        System.out.println("시도할 회수는 몇회인가요?");
-        return scanner.nextLine();
-    }
 
     public void printRoundResult(List<RacingCar> cars) {
         System.out.println("*** 실행결과 ***");
